feat: Support case filters in templates
This adds
- capitalize (stdlib)
- pascal_case (custom)
- kebab_case (custom)
- snake_case (custom)

Some might be duplicate of built-in behavior but that is in prep for
experimenting with #48.

We could add more case filters but they seemed unlikely to be used.

There are a lot of other string filters in liquid that could be useful
as well as a "slugify" (turn arbitrary strings into identifiers) or
"deunicode" which converts unicode characters to ascii.  For now, I'm
starting conservatively.

Fixes #117