Used vendored openssl for macosx builds
Travis CI's OSX environment causes the build to dynamically link to openssl@1.1 which isn't installed by default on mac machines. For consistency with the linux binary, we can use the vendored-openssl feature to sidestep this issue, which will eliminate having to install openssl@1.1 for users consuming the release binary. This increases the build size for macos by ~2mb.