Only set initial branch if given
Currently, `git::init` always sets the `initial_head` to the string
passed in with `--branch`, defaulting to `--master` if it's not given.
By not setting it unless `--branch` is explicitly given, we respect the
user's `config.defaultBranch`.

Note that this has the (possibly?) confusing behavior of copying the
branch name used for cloning the *template* repository into the
newly-created repository.