avoid using toml::Value in cargo-generate types if possible
Prior to this commit, `TemplateSlotsTable` was more of a
`TomlValueTable` which complicated things when it came to writing
`TemplateSlotsTable.is_valid` which if not strictly necessary, is
certainly useful as an abstraction in validating placeholder values.

As a general principle, I also find that it's good in a given library to
avoid defining one's types in terms of another library's type. I say
this because in theory `toml::Value` is only an implementation detail,
just one potential source of the cargo-generate `Config` struct.