Update dependencies
Update cargo.toml to use "~x.y" in the version specifiers, in order to lock it to that specific minor version... otherwise `cargo update` happily updated e.g. rhai from the specified 1.8 to the newest 1.12.

This actually caused cargo generate NOT TO BUILD sometimes...

Better to have slightly outdated dependecies that a broken build!