add --skip-submodules flag to optionalize cloning git submodules
`cargo-generate` is typically used for generating templates from small repos.
However, we would like to use it on our repo [1] where the template lies in a
larger mono-repo with git submodules that are not necessary to the template.
The template is kept in this repo to failitate testing the various components
within this repo. In this situation, the submodule download can add significat
overhead to the initialization of the repo. Some have reported running our
template generation command for tens of minutes and impacts the usability of
our tool.

This change makes submodule cloning optional by adding a --skip-submodules flag
that callers can set when using cargo-generate crate functions. This change
intends the default behavior to clone submodules if they exist.

[1]: https://github.com/risc0/risc0