feat: be less restrictive with `tempfile` dependency (#1591)
This causes issues when depending on the head of the
[`wasmtime`](https://github.com/bytecodealliance/wasmtime/tree/main)
repo, as it uses tempfile
[`3.23.0`](https://github.com/bytecodealliance/wasmtime/blob/main/Cargo.toml#L390)
(the latest version), but `cargo-generate` conflicts with its choice of
`~3.20.0`, which only allows resolving with a different patch version
(instead of minor), resulting in the following error:

```sh
/dev/mono/src/rust/test-tempfile> cargo check
    Blocking waiting for file lock on package cache
    Updating git repository `https://github.com/bytecodealliance/wasmtime`
    Updating git submodule `https://github.com/WebAssembly/component-model`
    Updating git submodule `https://github.com/WebAssembly/testsuite`
    Updating git submodule `https://github.com/WebAssembly/wasi-testsuite.git`
    Updating git submodule `https://github.com/WebAssembly/wasi-threads`
    Updating crates.io index
error: failed to select a version for `tempfile`.
    ... required by package `wasmtime v40.0.0 (https://github.com/bytecodealliance/wasmtime#0b9ff9bf)`
    ... which satisfies git dependency `wasmtime` of package `test-tempfile v0.1.0 (/Users/theo/dev/mono/src/rust/test-tempfile)`
versions that meet the requirements `^3.23.0` are: 3.23.0

all possible versions conflict with previously selected packages.

  previously selected package `tempfile v3.20.0`
    ... which satisfies dependency `tempfile = "~3.20"` (locked to 3.20.0) of package `cargo-generate v0.23.6`
    ... which satisfies dependency `cargo-generate = "^0.23.6"` (locked to 0.23.6) of package `test-tempfile v0.1.0 (/Users/theo/dev/mono/src/rust/test-tempfile)`

failed to select a version for `tempfile` which could resolve this conflict
```

You can reproduce this with a `Cargo.toml` like so:
```toml
[package]
name = "test-tempfile"
version = "0.1.0"
edition = "2024"

[dependencies]
cargo-generate = "0.23.6"
wasmtime = { git = "https://github.com/bytecodealliance/wasmtime" }
```

The cargo book [generally recommends caret
requirements](https://doc.rust-lang.org/nightly/cargo/reference/resolver.html#recommendations)
for most situations, so this change seems like a fine choice to me 🤷